Generalize clickable items to "ingredients"
Recipes can include sources or products that aren't items. For example,
NPC drops essentially have the NPC as an input and items as outputs. So
I've generalized them to just be "ingredients". In the future, this
can be expanded to include NPCs, tiles, and even more exotic things like
biomes and liquids.