Simplify recipe handler design
Previously, separate source and usage handlers would have to be used,
even if they were conceptually the same tab. Now, the distinction is
made by passing a `QueryType` parameter to `GetRecipeDisplays`, which
can be used to ask for either sources or uses. It is also much simpler
to store a `QueryType` in a history entry than to store a pointer to a
list of tabs.