Actually fix the autopause input problem
Recipe Browser's method of calling `ProcessTriggers` manually results in
some weird situations where it's called multiple times per update. So
instead, I've just copied the conditions for tModLoader's
`ProcessTriggers`, and handled input in `ModSystem::UIUpdate` instead.
It doesn't *seem* to be activated when typing in the search bar or
anything, so I guess it works?